requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
streamlit>=1.30.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
